Move `run_renderer` into the main `global_ctxt` closure
